% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{methods}
\name{as_tibble,DNAStringSet-method}
\alias{as_tibble,DNAStringSet-method}
\title{as_tibble for DNAStringSet}
\usage{
\S4method{as_tibble}{DNAStringSet}(x)
}
\description{
as_tibble for DNAStringSet
}
