% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{methods}
\name{as_tibble,otu_table-method}
\alias{as_tibble,otu_table-method}
\title{as_tibble for otu_table}
\usage{
\S4method{as_tibble}{otu_table}(x, tidy = FALSE)
}
\description{
If tidy == TRUE, will give in tidy format like psmelt
}
