% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{methods}
\name{as_tibble,phyloseq-method}
\alias{as_tibble,phyloseq-method}
\title{as_tibble for phyloseq objects}
\usage{
\S4method{as_tibble}{phyloseq}(x, tax = TRUE)
}
\arguments{
\item{tax}{whether to include taxonomy information}
}
\description{
as_tibble for phyloseq objects
}
