% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{methods}
\name{as_tibble,sample_data-method}
\alias{as_tibble,sample_data-method}
\title{as_tibble for sample_data}
\usage{
\S4method{as_tibble}{sample_data}(x)
}
\description{
as_tibble for sample_data
}
