% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\docType{methods}
\name{as_tibble,taxonomyTable-method}
\alias{as_tibble,taxonomyTable-method}
\title{as_tibble for taxonomyTable}
\usage{
\S4method{as_tibble}{taxonomyTable}(x)
}
\description{
as_tibble for taxonomyTable
}
