% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-methods.R
\name{as_tibble.otu_table}
\alias{as_tibble.otu_table}
\title{as_tibble for otu_table}
\usage{
as_tibble.otu_table(x, tidy = FALSE)
}
\description{
If tidy == TRUE, will give in tidy format like psmelt
}
