% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-methods.R
\name{as_tibble.phyloseq}
\alias{as_tibble.phyloseq}
\title{as_tibble for phyloseq objects}
\usage{
as_tibble.phyloseq(x, tax = TRUE)
}
\arguments{
\item{tax}{whether to include taxonomy information}
}
\description{
as_tibble for phyloseq objects
}
