% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{diversity}
\alias{diversity}
\title{True diversity of composition x (i.e., Hill numbers)}
\usage{
diversity(x, order = 1)
}
\description{
Diversity sensu Jost (i.e., Hill numbers, effective number of species, true
diversity; Jost2006)
}
