% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-filter.R
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa2}
\title{Filter taxa based on across-sample OTU abundance criteria}
\usage{
filter_taxa(physeq, .f, prune = FALSE)

filter_taxa2(physeq, .f)
}
\description{
Wrappers around phyloseq::filter_taxa that allow using a purrr-style
anonymous function for \code{.f}.
}
\section{Functions}{
\itemize{
\item \code{filter_taxa2}: Sets \code{prune = TRUE}
}}

\examples{
library(phyloseq)
library(tidyseq)
data(GlobalPatterns)
# Filter low prevalence taxa and then convert to proportions
gp.prop <- GlobalPatterns \%>\%
  filter_taxa2(~ sum(. > 0) > 5) \%>\%
  transform_sample_counts(~ . / sum(.))
}
\seealso{
phyloseq::filter_taxa
tidyseq::transform_sample_counts
}
