% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-constructors.R
\docType{methods}
\name{tax_table,tbl_df-method}
\alias{tax_table,tbl_df-method}
\title{Constructor for creating taxonomyTable from a tibble (class tbl_df)}
\usage{
\S4method{tax_table}{tbl_df}(object)
}
\description{
Assumes that the first column is the OTU/taxon names that will become
\code{taxa_names(physeq)}.
}
\details{
Note: Unlike phyloseq's method for data.frames, this does not coerce the
columns to character vectors; but will probably change this.
}
