% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-filter.R
\name{transform_sample_counts}
\alias{transform_sample_counts}
\title{Transform abundance data in an \code{otu_table}, sample-by-sample.}
\usage{
transform_sample_counts(physeq, .f, ...)
}
\description{
This is a wrapper around phyloseq::transform_sample_counts that allows use
of a purrr-style anonymous function for \code{.f}.
}
\examples{
library(phyloseq)
library(tidyseq)
data(GlobalPatterns)
# Filter low prevalence taxa and then convert to proportions
gp.prop <- GlobalPatterns \%>\%
  filter_taxa(~ sum(. > 0) > 5) \%>\%
  transform_sample_counts(~ . / sum(.))
}
\seealso{
phyloseq::transform_sample_counts
tidyseq::filter_taxa
}
