% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-filter.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter taxa based on across-sample OTU abundance criteria}
\usage{
filter_taxa(physeq, .f, prune = TRUE)
}
\description{
This is a wrapper around phyloseq::filter_taxa. It allows use of a
purrr-style anonymous function for \code{.f}. It also sets prune = TRUE by
default to be consistent with other subsetting functions and facilitate use
in pipe chains.
}
\examples{
library(phyloseq)
library(tidyseq)
data(GlobalPatterns)
# Filter low prevalence taxa and then convert to proportions
gp.prop <- GlobalPatterns \%>\%
  filter_taxa(~ sum(. > 0) > 5) \%>\%
  transform_sample_counts(~ . / sum(.))
}
\seealso{
phyloseq::filter_taxa
tidyseq::transform_sample_counts
}
