% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-constructors.R
\docType{methods}
\name{otu_table,tbl_df-method}
\alias{otu_table,tbl_df-method}
\title{Constructor for creating otu_table from a tibble (class tbl_df)}
\usage{
\S4method{otu_table}{tbl_df}(object, taxa_are_rows)
}
\description{
Assumes that the first column is the OTU/taxon names or the sample names,
depending on taxa_are_rows
}
