% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-constructors.R
\docType{methods}
\name{sample_data,tbl_df-method}
\alias{sample_data,tbl_df-method}
\title{Constructor for creating sample_data from a tibble (class tbl_df)}
\usage{
\S4method{sample_data}{tbl_df}(object)
}
\description{
Assumes that the first column is the sample names that will become
\code{sample_names(physeq)}
}
\details{
Converts to a plain data.frame with row names = sample names and then passes
to phyloseq's sample_data
}
