% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq-constructors.R
\docType{methods}
\name{tax_table,tbl_df-method}
\alias{tax_table,tbl_df-method}
\title{Constructor for creating taxonomyTable from a tibble (class tbl_df)}
\usage{
\S4method{tax_table}{tbl_df}(object)
}
\description{
Assumes that the first column is the OTU/taxon names.
}
\details{
This function creates a matrix from \code{object} that contains all columns
except the first one, which is used for the row names. The resulting matrix
is then passed to \code{tax_table()}.
}
\examples{
\dontrun{
# Read a .csv file with readr, which creates an object of class `tbl_df`
tbl <- readr::read_csv("path/to/taxonomy_table.csv")
# Inspect and check that the first column contains the taxa/OTU names
head(tbl) 
# Create a phyloseq `taxonomyTable` object
tax <- tax_table(tbl)
}
}
\seealso{
phyloseq::tax_table
}
